+++ src/tiled/flipmapobjects.h
@@ -1,6 +1,6 @@
 /*
  * flipmapobjects.h
- * Copyright 2013, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
+ * Copyright 2013-2022, Thorbjørn Lindeijer <thorbjorn@lindeijer.nl>
  *
  * This file is part of Tiled.
  *
@@ -20,13 +20,14 @@
 
 #pragma once
 
-#include "mapobject.h"
+#include "tiled.h"
 
 #include <QList>
 #include <QUndoCommand>
 
 namespace Tiled {
 
+class MapObject;
 class Document;
 
 class FlipMapObjects : public QUndoCommand
@@ -34,7 +35,8 @@ class FlipMapObjects : public QUndoCommand
 public:
     FlipMapObjects(Document *document,
                    const QList<MapObject *> &mapObjects,
-                   FlipDirection flipDirection);
+                   FlipDirection flipDirection,
+                   QPointF flipOrigin);
 
     void undo() override { flip(); }
     void redo() override { flip(); }
@@ -44,8 +46,8 @@ private:
 
     Document *mDocument;
     const QList<MapObject *> mMapObjects;
-    QPointF mObjectsCenter;
     FlipDirection mFlipDirection;
+    QPointF mFlipOrigin;
 
     QVector<bool> mOldRotationStates;
     QVector<bool> mNewRotationStates;